Port to Fabric MC 1.19.3

+++ src/main/java/me/jddev0/ep/energy/EnergyStoragePacketUpdate.java
@@ -4,6 +4,6 @@ package me.jddev0.ep.energy;
  * Used for EnergySyncS2CPacket
  */
 public interface EnergyStoragePacketUpdate {
-    void setEnergy(int energy);
-    void setCapacity(int capacity);
+    void setEnergy(long energy);
+    void setCapacity(long capacity);
 }
